requirements.txt
# =============================================================================
# Final Expense Voice Bot — Python Dependencies
# =============================================================================
#
# Install: pip install -r requirements.txt
#
# For GPU servers, also install:
#   pip install nemo_toolkit[asr] torch torchaudio soundfile
# =============================================================================

# --- Web framework & async ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiohttp>=3.10.0
pyVoIP>=1.6.8

# --- Configuration ---
python-dotenv>=1.0.0
pyyaml>=6.0.0

# --- Data processing ---
numpy>=1.26.0

# --- Type checking / dev ---
pydantic>=2.9.0

# --- Audio ---
soundfile>=0.12.0

# --- Audio I/O for voice pipeline test (test_voice_pipeline.py) ---
# Install manually as needed:
#   pip install sounddevice        — mic recording + speaker playback
#   pip install faster-whisper     — CPU STT (Windows-friendly alternative to Parakeet)
# sounddevice>=0.5.0
# faster-whisper>=1.0.0

# --- NeMo ASR (Parakeet TDT) — install separately on GPU servers ---
# nemo_toolkit[asr]>=2.0.0
# torch>=2.4.0
# torchaudio>=2.4.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.24.0
